backend/requirements.txt
fastapi==0.110.0
SQLAlchemy==1.4.52
alembic==1.13.1
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncio==3.4.3
pydantic-settings==2.2.1
email-validator==2.3.0
bcrypt==3.2.0
